% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_gus.R
\name{retrieve_gus}
\alias{retrieve_gus}
\title{Retrieve data for Pilica Catchment NUTS6 units}
\usage{
retrieve_gus(topic, save = TRUE, verbose = TRUE)
}
\arguments{
\item{topic}{{string selected from the list of following topics:
\itemize{
\item forests
\item inhabitants
\item companies_Ikw_0_9
\item companies_Ikw_10_49
\item companies_Ikw_50_249
\item companies_Ikw_250_999
\item companies_Ikw_1000
\item companies_Ikw_total
\item farms_mineral_2010
\item farms_wapn_2010
\item farms_potas_2010
\item farms_azot_2010
\item farms_fosfor_2010
\item treatment_plants349
\item treatment_plants348
\item waterworks
\item sewers
\item farms_cattle_cows_2010
\item farms_cattle_pigs_2010
\item farms_cattle_chick_2010
\item farms_cattle_hors_2010
\item farms_total_ha
\item farms_to_1ha
\item farms_over_1ha
\item farms_1_5_ha
\item farms_5_10_ha
\item farms_10_15_ha}
}}

\item{save}{{TRUE or FALSE}}
}
\value{
Dataframe including retrieved data for Pilica Catchment NUTS6 units for selected "topic". If \code{save = TRUE} the file \code{topic.rds} in \code{retrieveddata} directory is created.
If data for some NUTS6 have been not retrieved successfully, their names appear in message after function execution.

@examples
retrieve_gus("sewers")
}
\description{
{Retrieve data for 89 Pilica Catchment NUTS6 units from Statistics Poland (GUS BDL) database through GUS BDL API}
}
