% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_gus.R
\name{transform_gus}
\alias{transform_gus}
\title{Integrate datasets retrieved with 'retrieve_gus' function into one dataframe}
\usage{
transform_gus()
}
\value{
Dataframe in "transformed".rds format file in \code{retrieveddata} directory. Dataframe includes data from all topic.rds type datasets stored in \code{retrieveddata} directory.
}
\description{
Integrate datasets retrieved with 'retrieve_gus' function into one dataframe
}
\examples{
transform_gus()

}
