% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_gus.R
\name{transform_gus}
\alias{transform_gus}
\title{Integrate datasets retrieved with 'retrieve_gus' function into one dataframe}
\usage{
transform_gus(save = TRUE)
}
\arguments{
\item{save}{{TRUE or FALSE; default = TRUE - save retrieved data into .rds file into \code{transformeddata} directory}}
}
\value{
Dataframe including data from all topic.rds type datasets stored in \code{retrieveddata} directory.
If \code{save = TRUE} the file \code{transformeddata.rds} in \code{transformeddata} directory is created.
}
\description{
Integrate datasets retrieved with 'retrieve_gus' function into one dataframe
}
\examples{
## Not run:
transform_gus()
## End(Not run)

}
